python/requirements.txt
# CityFlow Web — Python Services
# Dependencias para los microservicios de IA y visión artificial

# ── Framework web ──────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Agente IA (agno + Gemini) ──────────────────
agno>=2.5.0
google-genai>=0.5.0

# ── Búsqueda web (DuckDuckGo) ──────────────────
ddgs>=0.1.0

# ── Conexión Azure SQL ─────────────────────────
pyodbc>=5.1.0
python-dotenv>=1.0.0
requests>=2.31.0
